macro mix bugs (mantis: 324, 325, 326)

+++ far/filelist.hpp
@@ -218,7 +218,7 @@ class FileList:public Panel
     virtual int ProcessKey(int Key);
     virtual int ProcessMouse(MOUSE_EVENT_RECORD *MouseEvent);
     virtual void MoveToMouse(MOUSE_EVENT_RECORD *MouseEvent);
-    virtual int VMProcess(int OpCode,void *vParam=NULL,__int64 iParam=0);
+    virtual __int64 VMProcess(int OpCode,void *vParam=NULL,__int64 iParam=0);
     virtual void SetFocus();
     virtual void Update(int Mode);
     /*$ 22.06.2001 SKV
